实现倒计时逻辑。借助死信队列实现倒计时

+++ meditation-service/meditation-service/src/main/java/com/donffroodus/meditation_service/entity/MeditationRecord.java
@@ -25,8 +25,8 @@ public class MeditationRecord {
     // 我们定义一个简单的 String 或 Enum 来存储状态
     private String status; 
     
-    private Integer musicId;
-    private Integer imageId;
+    private Long musicId;
+    private Long imageId;
 
     @Column(insertable = false, updatable = false)
     private LocalDateTime createdAt;
